// frontend/src/types/generated/SafeAsset.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type SafeAsset = { id: string, originalFileName: string | null, type: string, fileCreatedAt: string | null, width: number | null, height: number | null, uploaderName: string | null, uploaderIsFallback: boolean, downloadUrl: string | null, };